+++ src/components/NoticeCard/index.tsx
@@ -1,31 +1,11 @@
 'use client';
 
 import * as S from './style';
+import { parseDateString } from '@/utils';
 
 import type { CardInfo } from '@/types';
 
 const NoticeCard: React.FC<CardInfo> = ({ name, date, content, tag }) => {
-  const parseDateString = (
-    dateString: string
-  ): { monthDay: string; time: string } => {
-    const date = new Date(dateString);
-
-    const monthDayOptions: Intl.DateTimeFormatOptions = {
-      month: 'long',
-      day: 'numeric',
-    };
-    const monthDay = date.toLocaleString('ko-KR', monthDayOptions);
-
-    const timeOptions: Intl.DateTimeFormatOptions = {
-      hour: 'numeric',
-      minute: 'numeric',
-      hour12: true,
-    };
-    const time = date.toLocaleString('ko-KR', timeOptions);
-
-    return { monthDay, time };
-  };
-
   const { monthDay, time } = parseDateString(date);
 
   return (
