+++ src/utils/index.ts
@@ -10,3 +10,4 @@ export { default as isExistCookie } from './isExistCookie';
 export { default as isNumeric } from './isNumeric';
 export { default as minutesToMs } from './minutesToMs';
 export { default as minutesToSeconds } from './minutesToSeconds';
+export { default as parseDateString } from './dateFormat';
